Move "Using cove" section to the top
It is fairly important and easily missed if it is placed after the long
"Manual install" section. If I ever add easier ways to install cove
(like providing prebuilt binaries or packaging it on some package
managers), I might change this order again and refer to the "Using cove"
section in the top paragraph instead.