Fix list cursor when item moves off-screen
When a list scrolls or changes in such a way that the cursor item moves
off-screen, the cursor would jump to the closest visible item.

It makes more sense for the cursor to remain on its selected item and
for the list to scroll instead. That way, it is less likely for the user
to perform an action on the wrong list item if they press a key while
the list is changing.